+++ examples/resize.rs
@@ -1,11 +1,12 @@
-use bevy_app::{App, EventReader};
+use bevy_app::{App, Update};
 use bevy_doryen::doryen::{AppOptions, TextAlign, DEFAULT_CONSOLE_HEIGHT, DEFAULT_CONSOLE_WIDTH};
 use bevy_doryen::{
-    DoryenPlugin, DoryenPluginSettings, Input, RenderSystemExtensions, ResizeMode, Resized,
-    RootConsole,
+    DoryenPlugin, DoryenPluginSettings, Input, Render, ResizeMode, Resized, RootConsole,
 };
-use bevy_ecs::system::{IntoSystem, Res, ResMut};
+use bevy_ecs::prelude::EventReader;
+use bevy_ecs::system::{Res, ResMut, Resource};
 
+#[derive(Resource)]
 struct ResizeData {
     width: u32,
     height: u32,
@@ -13,7 +14,7 @@ struct ResizeData {
 }
 
 fn main() {
-    App::build()
+    App::new()
         .insert_resource(DoryenPluginSettings {
             app_options: AppOptions {
                 window_title: String::from("resizable console"),
@@ -23,15 +24,14 @@ fn main() {
             resize_mode: ResizeMode::Callback(resize_callback),
             ..Default::default()
         })
-        .add_plugin(DoryenPlugin)
+        .add_plugins(DoryenPlugin)
         .insert_resource(ResizeData {
             width: DEFAULT_CONSOLE_WIDTH,
             height: DEFAULT_CONSOLE_HEIGHT,
             mouse_pos: (0.0, 0.0),
         })
-        .add_system(update_mouse_position.system())
-        .add_system(resize_events.system())
-        .add_doryen_render_system(render.system())
+        .add_systems(Update, (update_mouse_position, resize_events))
+        .add_systems(Render, render)
         .run();
 }
 
@@ -43,11 +43,7 @@ fn resize_callback(root_console: &mut RootConsole, resized: Resized) {
     root_console.resize(resized.new_width / 8, resized.new_height / 8)
 }
 
-fn resize_events(
-    mut resize_data: ResMut<ResizeData>,
-    //events: Res<Events<Resized>>,
-    mut event_reader: EventReader<Resized>,
-) {
+fn resize_events(mut resize_data: ResMut<ResizeData>, mut event_reader: EventReader<Resized>) {
     if let Some(resized) = event_reader.iter().last() {
         resize_data.width = resized.new_width / 8;
         resize_data.height = resized.new_height / 8;