+++ examples/text_input.rs
@@ -1,27 +1,19 @@
-use bevy_app::App;
+use bevy_app::{App, Update};
 use bevy_doryen::doryen::{AppOptions, Color, TextAlign};
-use bevy_doryen::{DoryenPlugin, DoryenPluginSettings, Input, RenderSystemExtensions, RootConsole};
-use bevy_ecs::system::{IntoSystem, Res, ResMut};
+use bevy_doryen::{DoryenPlugin, DoryenPluginSettings, Input, Render, RootConsole};
+use bevy_ecs::system::{Res, ResMut, Resource};
 use unicode_segmentation::UnicodeSegmentation;
 
 const WHITE: Color = (255, 255, 255, 255);
 
+#[derive(Default, Resource)]
 struct TextInput {
     text: String,
     cursor: usize,
 }
 
-impl Default for TextInput {
-    fn default() -> Self {
-        Self {
-            text: String::new(),
-            cursor: 0,
-        }
-    }
-}
-
 fn main() {
-    App::build()
+    App::new()
         .insert_resource(DoryenPluginSettings {
             app_options: AppOptions {
                 window_title: String::from("bevy_doryen subcell resolution demo"),
@@ -29,10 +21,10 @@ fn main() {
             },
             ..Default::default()
         })
-        .add_plugin(DoryenPlugin)
+        .add_plugins(DoryenPlugin)
         .init_resource::<TextInput>()
-        .add_system(update.system())
-        .add_doryen_render_system(render.system())
+        .add_systems(Update, update)
+        .add_systems(Render, render)
         .run();
 }
 
@@ -40,7 +32,7 @@ fn update(input: Res<Input>, mut text_input: ResMut<TextInput>) {
     // input.text returns the characters typed by the player since last update
     let text = input.text();
     if !text.is_empty() {
-        text_input.text.push_str(&text);
+        text_input.text.push_str(text);
     }
     // handle backspace
     if input.key_released("Backspace") && !text_input.text.is_empty() {